requirements.txt
fastapi
uvicorn
pydantic
google-genai
youtube-transcript-api